feature(Settings): Refactoring code, splitting out widgets module, fixing style issues
