+++ src/containers/SettingsEditor/Fields/FormFields.styled.ts
@@ -1,0 +1,18 @@
+import styled from 'styled-components'
+
+export const ArrayItemControls = styled.div`
+  display: flex;
+  flex-direction: row;
+  justify-content: flex-start;
+  gap: 5px;
+
+  button {
+    border-radius: 50%;
+    width: 30px;
+    height: 30px;
+  }
+`
+
+export const ItemWrapper = styled.div`
+  margin-bottom: 8px;
+`
